Toggle pings, remove line-wrapping, unset victim.
There's a new option (-P) to hide pings at startup, and a new command (/P)
to toggle ping visibility.  Pings are sometimes useful to know about but
if you don't have a target, they kinda flood.

On startup, target is nil and this lets you see everything coming through.
I'd miss DMs because the target was set to a channel, so I had to annoy
people by telling them that I couldn't see DMs unless they told me that I had
them before I had one.  So I added this option and then got a screen full of
PINGs when I woke up.  (Or half full of PINGs and the balance in PONGs.)
This makes irc7 usable as a general client for me.

Lastly, I removed line-wrapping.  I have no use for it.  If anyone else cares
about this changeset, then maybe it could be an option, but as long as I'm the
only one using it, I don't need an option that I'll never enable.

khm's C code is a fun read.  It has a lot of personality.